requirements.txt
pandas
requests
sqlalchemy
tenacity
python-dotenv
psycopg2-binary
plotly
streamlit